forgot password flow with mock API and navigation refractoring
