+++ Blood_Management_App/.flowconfig
@@ -70,4 +70,4 @@ untyped-import
 untyped-type-import
 
 [version]
-^0.122.0
+^>=0.122.0
