+++ Blood_Management_App/constants/Colors.js
@@ -1,6 +1,7 @@
 /* eslint-disable prettier/prettier */
 const colors = {
   primary: '#E94364',
+  darkPrimary: '#b02e47',
   secondary: '#58ACE0',
   accent: '#cccccc',
   additional1: 'black',
