+++ Blood_Management_App/redux/RootReducer.js
@@ -5,12 +5,14 @@ import loginReducer from './login/reducer';
 import regReducer from './register/reducer';
 import regIndReducer from './registerInd/reducer';
 import authReducer from './auth/reducer';
+import forgotReducer from './forgotpassword/reducer';
 
 const rootReducer = combineReducers({
   loginFormState: loginReducer,
   regFormState: regReducer,
   regIndFormState: regIndReducer,
   authState: authReducer,
+  forgotState: forgotReducer,
 });
 
 export default rootReducer;
