+++ Blood_Management_App/redux/forgotpassword/actionTypes.js
@@ -1,0 +1,7 @@
+export const UPDATE_FIELDS_FORGOT = 'UPDATE_FIELDS_FORGOT';
+export const FORGOT_REQ = 'FORGOT_REQ';
+export const FORGOT_REQ_SUCCESS = 'FORGOT_REQ_SUCCESS';
+export const FORGOT_REQ_FAILURE = 'FORGOT_REQ_FAILURE';
+export const BLUR_FIELDS_FORGOT = 'BLUR_FIELDS_FORGOT';
+export const STATE_CLEANUP_FORGOT = 'STATE_CLEANUP_FORGOT';
+export const RESET_DONE_STATE = 'RESET_DONE_STATE';
