--- Blood_Management_App/src/MainStack/Individual/Services.js
@@ -1,22 +1,0 @@
-import React from 'react'
-import {
-    View,
-    Text,
-    TouchableOpacity
-
-} from 'react-native'
-
-
-const Services = ({navigation}) => {
-
-    return(
-        <View style={{flex:1,justifyContent:"center", alignContent:'center',  alignItems:'center'}}>
-            <Text>
-            Services
-            </Text>
-           
-        </View>
-    );
-}
-
-export default Services
