+++ Blood_Management_App/src/MainTabs/About.js
