+++ Blood_Management_App/src/MainTabs/Individual/Home.js
