+++ Blood_Management_App/src/MainTabs/ServicesStack/Services.js
@@ -1,0 +1,22 @@
+/* eslint-disable prettier/prettier */
+import React from 'react';
+import {View, Text, StyleSheet} from 'react-native';
+
+const Services = ({navigation}) => {
+  return (
+    <View style={styles.container}>
+      <Text>Services</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignContent: 'center',
+    alignItems: 'center',
+  },
+});
+
+export default Services;
